requirements.txt
fastapi==0.123.0
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.22
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pydantic==2.12.5
python-multipart==0.0.20
aiofiles==25.1.0
supabase== 2.24.0

httpx==0.28.1

